conditionalize MPI_UB/MPI_LB on existence
Open-MPI 4 implements the MPI-3 standard deletion of MPI_UB and MPI_LB.
There is a backwards-compatibility mode, so we test for Open-MPI 4 and
not the compatibility mode when disabling references to these
identifiers.

There may be other valid strategies for dealing with this problem...

See https://www.open-mpi.org/faq/?category=mpi-removed#mpi-1-mpi-lb-ub
for details.

Signed-off-by: Jeff Hammond <jeff.r.hammond@intel.com>